+++ subprojects/jacoco/src/main/java/org/gradle/internal/jacoco/AntJacocoReport.java
@@ -17,72 +17,18 @@
 package org.gradle.internal.jacoco;
 
 import com.google.common.collect.ImmutableMap;
-import groovy.lang.Closure;
 import groovy.lang.GroovyObjectSupport;
-import org.gradle.api.file.FileCollection;
 import org.gradle.api.internal.project.IsolatedAntBuilder;
 import org.gradle.testing.jacoco.tasks.JacocoReportsContainer;
 
-import java.util.Collections;
-import java.util.Map;
-
-public class AntJacocoReport {
-
-    private final IsolatedAntBuilder ant;
+public class AntJacocoReport extends AbstractAntJacocoReport<JacocoReportsContainer> {
 
     public AntJacocoReport(IsolatedAntBuilder ant) {
-        this.ant = ant;
+        super(ant);
     }
 
-    public void execute(FileCollection classpath, final String projectName,
-                        final FileCollection allClassesDirs, final FileCollection allSourcesDirs,
-                        final FileCollection executionData,
-                        final JacocoReportsContainer reports) {
-        ant.withClasspath(classpath).execute(new Closure<Object>(this, this) {
-            @SuppressWarnings("UnusedDeclaration")
-            public Object doCall(Object it) {
-                final GroovyObjectSupport antBuilder = (GroovyObjectSupport) it;
-                antBuilder.invokeMethod("taskdef", ImmutableMap.of(
-                    "name", "jacocoReport",
-                    "classname", "org.jacoco.ant.ReportTask"
-                ));
-                final Map<String, Object> emptyArgs = Collections.<String, Object>emptyMap();
-                antBuilder.invokeMethod("jacocoReport", new Object[]{emptyArgs, new Closure<Object>(this, this) {
-                    public Object doCall(Object ignore) {
-                        antBuilder.invokeMethod("executiondata", new Object[]{emptyArgs, new Closure<Object>(this, this) {
-                            public Object doCall(Object ignore) {
-                                executionData.addToAntBuilder(antBuilder, "resources");
-                                return null;
-                            }
-                        }});
-                        Map<String, Object> structureArgs = ImmutableMap.<String, Object>of("name", projectName);
-                        antBuilder.invokeMethod("structure", new Object[]{structureArgs, new Closure<Object>(this, this) {
-                            public Object doCall(Object ignore) {
-                                antBuilder.invokeMethod("classfiles", new Object[]{emptyArgs, new Closure<Object>(this, this) {
-                                    public Object doCall(Object ignore) {
-                                        allClassesDirs.addToAntBuilder(antBuilder, "resources");
-                                        return null;
-                                    }
-                                }});
-                                antBuilder.invokeMethod("sourcefiles", new Object[]{emptyArgs, new Closure<Object>(this, this) {
-                                    public Object doCall(Object ignore) {
-                                        allSourcesDirs.addToAntBuilder(antBuilder, "resources");
-                                        return null;
-                                    }
-                                }});
-                                return null;
-                            }
-                        }});
-                        configureReportFormats(antBuilder, reports);
-                        return null;
-                    }
-                }});
-                return null;
-            }
-        });
-    }
-
-    private void configureReportFormats(GroovyObjectSupport antBuilder, JacocoReportsContainer reports) {
+    @Override
+    protected void configureReport(GroovyObjectSupport antBuilder, JacocoReportsContainer reports) {
         if (reports.getHtml().isEnabled()) {
             antBuilder.invokeMethod("html", new Object[]{
                 ImmutableMap.<String, Object>of("destdir", reports.getHtml().getDestination())
